relocate config related classes to own package
