+++ Shared/src/main/java/dev/tr7zw/itemswapper/config/ConfigManager.java
@@ -1,6 +1,7 @@
-package dev.tr7zw.itemswapper;
+package dev.tr7zw.itemswapper.config;
 
 import com.google.gson.JsonSyntaxException;
+import dev.tr7zw.itemswapper.ItemSwapperSharedMod;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
