+++ Shared/src/main/java/dev/tr7zw/itemswapper/mixin/PlayerListMixin.java
@@ -5,7 +5,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import dev.tr7zw.itemswapper.ConfigManager;
+import dev.tr7zw.itemswapper.config.ConfigManager;
 import dev.tr7zw.itemswapper.util.NetworkLogic;
 import net.minecraft.network.Connection;
 import net.minecraft.server.level.ServerPlayer;
